use thunk for axios calls
