+++ redux/types/action_types.ts
@@ -1,0 +1,47 @@
+import { iThread, iUser } from "../../components/models";
+
+export interface IAppActions {
+    user: IAppUserActions;
+    thread: IAppThreadActions;
+}
+
+export enum ThreadAction {
+    GET_ALL = 'Get All',
+    ADD_THREAD = 'Add Thread',
+    REMOVE_THREAD = 'Remove Thread',
+    EDIT_THREAD = 'Edit Thread',
+    SELECT_THREAD = 'Select Thread',
+}
+export interface IThreadActions {
+    type: ThreadAction;
+    payload: { thread: iThread }
+}
+export interface IAppThreadActions extends IThreadActions {
+    type: ThreadAction;
+    payload: {
+        thread: iThread,
+        selection: number
+    }
+}
+
+export enum UserAction {
+    ADD_USER = 'Add User',
+    REMOVE_USER = 'Remove User',
+    EDIT_USER = 'Edit User',
+    SELECT_USER = 'Select User',
+    LOGIN_USER = 'Login User',
+    LOGOUT_USER = 'Logout User,'
+}
+
+export interface IUserActions {
+    type: UserAction;
+    payload: { user: iUser }
+}
+
+export interface IAppUserActions extends IUserActions {
+    type: UserAction;
+    payload: {
+        user: iUser,
+        selection: number
+    }
+}
